Allow for multiple binaries
* Added `symlink` method to FileSystemUtilities
* Added `bin` property to Package
* Updated bootstrap to factor in dependencies with multiple binaries
* Added test for package with multiple binaries
* Refactored way ignored packages are determined (to be more like run command)
* Updated unit test for ignored package

When a package is ignored while bootstrapping, it should be completely taken out of the equation. This update alters the previous behavior of not bootstrapping an ignored package, but still treating it as though it was part of the process. If a package is ignored, it should be treated as an external dependency. The updated unit test reflects this.